+++ json_serializable/test/kitchen_sink/kitchen_sink.g_any_map__checked.dart
@@ -65,9 +65,14 @@ class _Factory implements k.KitchenSinkFactory<dynamic, dynamic> {
         [],
         BigInt.zero,
         {},
+        BigInt.zero,
+        {},
         TrivialNumber(0),
         {},
         DateTime.fromMillisecondsSinceEpoch(0),
+        TrivialString(''),
+        TrivialNumber(0),
+        {},
       );
 
   k.JsonConverterTestClass jsonConverterFromJson(Map<String, dynamic> json) =>
@@ -191,14 +196,14 @@ class KitchenSink implements k.KitchenSink {
   }
 }
 
-@JsonSerializable(
-  checked: true,
-  anyMap: true,
-)
+@JsonSerializable(checked: true, anyMap: true, converters: [
+  // referencing a top-level field should work
+  durationConverter,
+  // referencing via a const constructor should work
+  BigIntStringConverter(),
+])
 // referencing a top-level field should work
-@durationConverter
-// referencing via a const constructor should work
-@BigIntStringConverter()
+@trivialStringConverter
 @TrivialNumberConverter.instance
 @EpochDateTimeConverter()
 class JsonConverterTestClass implements k.JsonConverterTestClass {
@@ -207,9 +212,14 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
     this.durationList,
     this.bigInt,
     this.bigIntMap,
+    this.nullableBigInt,
+    this.nullableBigIntMap,
     this.numberSilly,
     this.numberSillySet,
     this.dateTime,
+    this.trivialString,
+    this.nullableNumberSilly,
+    this.nullableNumberSillySet,
   );
 
   factory JsonConverterTestClass.fromJson(Map<String, dynamic> json) =>
@@ -223,10 +233,18 @@ class JsonConverterTestClass implements k.JsonConverterTestClass {
   BigInt bigInt;
   Map<String, BigInt> bigIntMap;
 
+  BigInt? nullableBigInt;
+  Map<String, BigInt?> nullableBigIntMap;
+
   TrivialNumber numberSilly;
   Set<TrivialNumber> numberSillySet;
 
   DateTime? dateTime;
+
+  TrivialString? trivialString;
+
+  TrivialNumber? nullableNumberSilly;
+  Set<TrivialNumber?> nullableNumberSillySet;
 }
 
 @JsonSerializable(