+++ json_serializable/test/kitchen_sink/kitchen_sink.g_exclude_null.g.dart
@@ -145,20 +145,34 @@ Map<String, dynamic> _$KitchenSinkToJson(KitchenSink instance) {
 JsonConverterTestClass _$JsonConverterTestClassFromJson(
         Map<String, dynamic> json) =>
     JsonConverterTestClass(
-      durationConverter.fromJson(json['duration'] as int?),
+      const DurationMillisecondConverter().fromJson(json['duration'] as int?),
       (json['durationList'] as List<dynamic>)
-          .map((e) => durationConverter.fromJson(e as int?))
+          .map((e) => const DurationMillisecondConverter().fromJson(e as int?))
           .toList(),
       const BigIntStringConverter().fromJson(json['bigInt'] as String),
       (json['bigIntMap'] as Map<String, dynamic>).map(
         (k, e) =>
             MapEntry(k, const BigIntStringConverter().fromJson(e as String)),
       ),
+      _$JsonConverterFromJson<String, BigInt>(
+          json['nullableBigInt'], const BigIntStringConverter().fromJson),
+      (json['nullableBigIntMap'] as Map<String, dynamic>).map(
+        (k, e) => MapEntry(
+            k,
+            _$JsonConverterFromJson<String, BigInt>(
+                e, const BigIntStringConverter().fromJson)),
+      ),
       TrivialNumberConverter.instance.fromJson(json['numberSilly'] as int?),
       (json['numberSillySet'] as List<dynamic>)
           .map((e) => TrivialNumberConverter.instance.fromJson(e as int?))
           .toSet(),
       const EpochDateTimeConverter().fromJson(json['dateTime'] as int?),
+      trivialStringConverter.fromJson(json['trivialString'] as String?),
+      TrivialNumberConverter.instance
+          .fromJson(json['nullableNumberSilly'] as int?),
+      (json['nullableNumberSillySet'] as List<dynamic>)
+          .map((e) => TrivialNumberConverter.instance.fromJson(e as int?))
+          .toSet(),
     );
 
 Map<String, dynamic> _$JsonConverterTestClassToJson(
@@ -171,12 +185,22 @@ Map<String, dynamic> _$JsonConverterTestClassToJson(
     }
   }
 
-  writeNotNull('duration', durationConverter.toJson(instance.duration));
-  val['durationList'] =
-      instance.durationList.map(durationConverter.toJson).toList();
-  writeNotNull('bigInt', const BigIntStringConverter().toJson(instance.bigInt));
+  writeNotNull('duration',
+      const DurationMillisecondConverter().toJson(instance.duration));
+  val['durationList'] = instance.durationList
+      .map(const DurationMillisecondConverter().toJson)
+      .toList();
+  val['bigInt'] = const BigIntStringConverter().toJson(instance.bigInt);
   val['bigIntMap'] = instance.bigIntMap
       .map((k, e) => MapEntry(k, const BigIntStringConverter().toJson(e)));
+  writeNotNull(
+      'nullableBigInt',
+      _$JsonConverterToJson<String, BigInt>(
+          instance.nullableBigInt, const BigIntStringConverter().toJson));
+  val['nullableBigIntMap'] = instance.nullableBigIntMap.map((k, e) => MapEntry(
+      k,
+      _$JsonConverterToJson<String, BigInt>(
+          e, const BigIntStringConverter().toJson)));
   writeNotNull('numberSilly',
       TrivialNumberConverter.instance.toJson(instance.numberSilly));
   val['numberSillySet'] = instance.numberSillySet
@@ -184,8 +208,30 @@ Map<String, dynamic> _$JsonConverterTestClassToJson(
       .toList();
   writeNotNull(
       'dateTime', const EpochDateTimeConverter().toJson(instance.dateTime));
+  writeNotNull(
+      'trivialString', trivialStringConverter.toJson(instance.trivialString));
+  writeNotNull(
+      'nullableNumberSilly',
+      _$JsonConverterToJson<int?, TrivialNumber>(instance.nullableNumberSilly,
+          TrivialNumberConverter.instance.toJson));
+  val['nullableNumberSillySet'] = instance.nullableNumberSillySet
+      .map((e) => _$JsonConverterToJson<int?, TrivialNumber>(
+          e, TrivialNumberConverter.instance.toJson))
+      .toList();
   return val;
 }
+
+Value? _$JsonConverterFromJson<Json, Value>(
+  Object? json,
+  Value? Function(Json json) fromJson,
+) =>
+    json == null ? null : fromJson(json as Json);
+
+Json? _$JsonConverterToJson<Json, Value>(
+  Value? value,
+  Json? Function(Value value) toJson,
+) =>
+    value == null ? null : toJson(value);
 
 JsonConverterGeneric<S, T, U> _$JsonConverterGenericFromJson<S, T, U>(
         Map<String, dynamic> json) =>
@@ -201,19 +247,10 @@ JsonConverterGeneric<S, T, U> _$JsonConverterGenericFromJson<S, T, U>(
     );
 
 Map<String, dynamic> _$JsonConverterGenericToJson<S, T, U>(
-    JsonConverterGeneric<S, T, U> instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('item', GenericConverter<S>().toJson(instance.item));
-  val['itemList'] =
-      instance.itemList.map(GenericConverter<T>().toJson).toList();
-  val['itemMap'] = instance.itemMap
-      .map((k, e) => MapEntry(k, GenericConverter<U>().toJson(e)));
-  return val;
-}
+        JsonConverterGeneric<S, T, U> instance) =>
+    <String, dynamic>{
+      'item': GenericConverter<S>().toJson(instance.item),
+      'itemList': instance.itemList.map(GenericConverter<T>().toJson).toList(),
+      'itemMap': instance.itemMap
+          .map((k, e) => MapEntry(k, GenericConverter<U>().toJson(e))),
+    };