+++ json_serializable/test/shared_config.dart
@@ -7,8 +7,9 @@ import 'package:json_serializable/src/type_helpers/config_types.dart';
 
 final jsonSerializableFields = generatorConfigDefaultJson.keys.toList();
 
-final generatorConfigDefaultJson =
-    Map<String, dynamic>.unmodifiable(ClassConfig.defaults.toJson());
+final generatorConfigDefaultJson = Map<String, dynamic>.unmodifiable(
+  ClassConfig.defaults.toJsonSerializable().toJson(),
+);
 
 // #CHANGE WHEN UPDATING json_annotation
 final generatorConfigNonDefaultJson =
@@ -18,6 +19,8 @@ final generatorConfigNonDefaultJson =
   constructor: 'something',
   createFactory: false,
   createToJson: false,
+  createFieldMap: true,
+  createPerFieldToJson: true,
   disallowUnrecognizedKeys: true,
   explicitToJson: true,
   fieldRename: FieldRename.kebab,